requirements.txt
# Gmail API dependencies
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Google AI dependencies
google-generativeai

# HTML processing
html2text
